% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busca.R
\name{buscar}
\alias{buscar}
\title{Buscar}
\usage{
buscar(q, sd, ed, max_r = 100, wt = 1)
}
\arguments{
\item{q}{query}

\item{sd}{data de inicio no formado dd/mm/aaaa}

\item{ed}{data de fim no formato dd/mm/aaaa}

\item{max_r}{número máximo de registros}

\item{wt}{tempo de espera entre as consultas (importante p/ não sobrecarregar
o site/ ser bloqueado)}
}
\description{
Função auxiliar que baixa todas as páginas de uma busca
dependendo de um número máximo de registros.
}

