% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busca.R
\name{buscar_}
\alias{buscar_}
\title{Buscar_}
\usage{
buscar_(q, sd, ed, sr = 1)
}
\arguments{
\item{q}{query}

\item{sd}{data de inicio no formado dd/mm/aaaa}

\item{ed}{data de fim no formato dd/mm/aaaa}

\item{sr}{primeiro item da consulta (a consulta sempre retorna 25 itens)}
}
\description{
Função auxiliar da busca.
}

