% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noticia.R
\name{folha_noticias}
\alias{folha_noticias}
\title{Obter mais infos das notícias.}
\usage{
folha_noticias(urls)
}
\arguments{
\item{urls}{vetor de urls para os quais deseja-se obter mais informações.}
}
\description{
Obter mais infos das notícias.
}
\examples{
urls <- c(
 "http://www1.folha.uol.com.br/mundo/2016/11/1830974-americanos-insatisfeitos-protestam-contra-vitoria-de-trump-em-eleicoes.shtml",
 "http://f5.folha.uol.com.br/colunistas/tonygoes/2016/12/sao-paulo-precisa-descobrir-sua-propria-maneira-de-realizar-a-virada-cultural.shtml",
 "http://f5.folha.uol.com.br/celebridades/2016/12/por-que-a-atriz-mais-bem-paga-da-tv-americana-esta-sendo-processada-pelos-proprios-embrioes.shtml",
 "http://www.agora.uol.com.br/saopaulo/2016/12/1839811-mortes-em-marginais-ocorrem-com-motos-nas-pistas-locais.shtml"
 )
noticias <- folha_noticias(urls)

}

