% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noticia.R
\name{parse_agora}
\alias{parse_agora}
\title{Faz o parse para noticias com URL começando em:
http://www.agora.uol.com.br/}
\usage{
parse_agora(url)
}
\arguments{
\item{url}{url da noticia}
}
\description{
Faz o parse para noticias com URL começando em:
http://www.agora.uol.com.br/
}
\examples{
url <- "http://www.agora.uol.com.br/saopaulo/2016/12/1839811-mortes-em-marginais-ocorrem-com-motos-nas-pistas-locais.shtml"
noticia <- parse_agora(url)

}

