% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noticia.R
\name{parse_f5}
\alias{parse_f5}
\title{Faz parse das notícias com URL começando em:
http://f5.folha.uol.com.br/}
\usage{
parse_f5(url)
}
\arguments{
\item{url}{url da noticia}
}
\description{
Faz parse das notícias com URL começando em:
http://f5.folha.uol.com.br/
}
\examples{
url <- "http://f5.folha.uol.com.br/colunistas/tonygoes/2016/12/sao-paulo-precisa-descobrir-sua-propria-maneira-de-realizar-a-virada-cultural.shtml"
noticia <- parse_f5(url)

url <- "http://f5.folha.uol.com.br/celebridades/2016/12/por-que-a-atriz-mais-bem-paga-da-tv-americana-esta-sendo-processada-pelos-proprios-embrioes.shtml"
noticia <- parse_f5(url)

}

