% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noticia.R
\name{parse_folha}
\alias{parse_folha}
\title{Faz parse das notícias com url começando em:
http://www1.folha.uol.com.br/}
\usage{
parse_folha(url)
}
\arguments{
\item{url}{url da noticia}
}
\description{
Faz parse das notícias com url começando em:
http://www1.folha.uol.com.br/
}
\examples{
url <- "http://www1.folha.uol.com.br/mundo/2016/11/1830974-americanos-insatisfeitos-protestam-contra-vitoria-de-trump-em-eleicoes.shtml"
noticia <- parse_folha(url)

}

