% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noticia.R
\name{parse_noticia}
\alias{parse_noticia}
\title{Parse Noticia}
\usage{
parse_noticia(url)
}
\arguments{
\item{url}{url da notícia}
}
\description{
Faz parse de uma notícia. Escolhe a função adequada, dependendo da URL.
}
\examples{
url <- "http://www.agora.uol.com.br/saopaulo/2016/12/1839811-mortes-em-marginais-ocorrem-com-motos-nas-pistas-locais.shtml"
parse_noticia(url)

url <- "http://www1.folha.uol.com.br/mundo/2016/11/1830974-americanos-insatisfeitos-protestam-contra-vitoria-de-trump-em-eleicoes.shtml"
parse_noticia(url)

url <- "http://f5.folha.uol.com.br/colunistas/tonygoes/2016/12/sao-paulo-precisa-descobrir-sua-propria-maneira-de-realizar-a-virada-cultural.shtml"
noticia <- parse_noticia(url)

url <- "http://f5.folha.uol.com.br/celebridades/2016/12/por-que-a-atriz-mais-bem-paga-da-tv-americana-esta-sendo-processada-pelos-proprios-embrioes.shtml"
noticia <- parse_noticia(url)

}

